% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_david_files.R
\name{combine_david_files}
\alias{combine_david_files}
\title{Combine text files produced by DAVID analysis}
\usage{
combine_david_files(path_to_folder)
}
\arguments{
\item{path_to_folder}{A string that is the path to the folder that contains
all the DAVID text files. The folder must only contain DAVID text files that
the user wants to analyze. The files must be names such that they are in
chronological order of their cluster number.}
}
\value{
A list of data frames, one data frame for each cluster
}
\description{
This function reads and combines text files produced by DAVID's annotation
chart into a list of data frames, one data frame for each file (each cluster)
}
\examples{
\dontrun{
library(SeuratToGO)
david_file_path = system.file("extdata", "david", package = "SeuratToGO")
# this is where DAVID output files are stored in this package
# when using your own files, you must provide the file path to the folder
# where you saved them
combined_list <- combine_david_files(david_file_path)
View(combined_list)
# see the vignette for more details
}

}
\references{
Sherman BT, Hao M, Qiu J, Jiao X, Baseler MW, Lane HC, Imamichi T, Chang W.
2022. DAVID: a web server for functional enrichment analysis and functional
annotation of gene lists (2021 update). \emph{Nucleic Acids Res}.
50(W1):W216–W221. \href{https://pubmed.ncbi.nlm.nih.gov/35325185/}{Link}
}
