% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_processes.R
\name{get_top_processes}
\alias{get_top_processes}
\title{Get the top GO processes in a cluster}
\usage{
get_top_processes(combined_list, cluster, benjamini = 0.05, top_n)
}
\arguments{
\item{combined_list}{A list of data frames (preferably the one produced by
combine_david_files function), where each data frame is a cluster, in
chronological order. This assumes the first data frame is cluster 0, the
second data frame is cluster 1, and so on.}

\item{cluster}{An integer that represents the cluster to be analyzed. For
example, 0 for cluster 0, 1 for cluster 1, and so on.}

\item{benjamini}{A number that represents the Benjamini threshold for
filtering processes. Only processes with a Benjamini value smaller than the
threshold will be kept. The default value is 0.05.}

\item{top_n}{The number of processes to return.}
}
\value{
A data frame with the top_n top processes.
}
\description{
This function find the top GO processes identified in a cluster, based on
the Benjamini threshold given.
}
