% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SeuratToGO.R
\name{run_SeuratToGO}
\alias{run_SeuratToGO}
\title{Run the package's Shiny App}
\usage{
run_SeuratToGO()
}
\description{
This function runs the package's Shiny App
}
\examples{
# SeuratToGO::run_SeuratToGO()


}
