% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_clusters.R
\name{separate_clusters}
\alias{separate_clusters}
\title{Separate genes based on cluster}
\usage{
separate_clusters(markers)
}
\arguments{
\item{markers}{A data frame of gene markers with columns "cluster" and "gene".}
}
\value{
A data frame of gene markets with cluster number as column name.
}
\description{
This function separates the genes based on their cluster, and labels the
the column name as the cluster number. It also exports the data frame as a
tab-delimited text file so it can be uploaded onto the DAVID Website.
}
