% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_processes_heatmap.R
\name{top_processes_heatmap}
\alias{top_processes_heatmap}
\title{Produce a heatmap of the top processes}
\usage{
top_processes_heatmap(top_processes_df, cellwidth = 30)
}
\arguments{
\item{top_processes_df}{A data frame containing the top processes of each
cluster. The rownames should be the processes and the column names should
be the cluster. It can be produced using the get_all_top_processes function
in the package.}

\item{cellwidth}{The cell width for each cell in the heatmap.}
}
\value{
A heatmap that represents the top processes in each cluster.
}
\description{
This function creates and returns a heatmap showing the top processes in each
cluster.
}
