% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_david_files.R
\name{combine_david_files}
\alias{combine_david_files}
\title{Combine text files produced by DAVID analysis}
\usage{
combine_david_files(path_to_folder)
}
\arguments{
\item{path_to_folder}{A string that is the path to the folder that contains
all the DAVID text files. The folder must only contain DAVID text files that
the user wants to analyze. The files must be names such that they are in
chronological order of their cluster number.}
}
\value{
A list of data frames, one data frame for each cluster
}
\description{
This function reads and combines text files produced by DAVID's annotation
chart into a list of data frames, one data frame for each file (each cluster)
}
