% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_top_processes.R
\name{get_all_top_processes}
\alias{get_all_top_processes}
\title{Get the p-value for the top processes of all clusters}
\usage{
get_all_top_processes(combined_list, benjamini, top_n)
}
\arguments{
\item{combined_list}{A list of data frames (preferably the one produced by
combine_david_files function), where each data frame is a cluster, in
chronological order. This assumes the first data frame is cluster 0, the
second data frame is cluster 1, and so on.}

\item{benjamini}{A number that represents the Benjamini threshold for
filtering processes. Only processes with a Benjamini value smaller than the
threshold will be kept. The default value is 0.05.}

\item{top_n}{The number of processes to identify for each cluster.}
}
\value{
A data frame with the top_n top processes for each cluster, combined.
}
\description{
This function finds the top processes identified for each cluster and compile
them into one data frame.
}
