% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_processes.R
\name{get_top_processes}
\alias{get_top_processes}
\title{Get the top GO processes in a cluster}
\usage{
get_top_processes(combined_list, cluster, benjamini = 0.05, top_n)
}
\arguments{
\item{combined_list}{A list of data frames (preferably the one produced by
combine_david_files function), where each data frame is a cluster, in
chronological order. This assumes the first data frame is cluster 0, the
second data frame is cluster 1, and so on.}

\item{cluster}{An integer that represents the cluster to be analyzed. For
example, 0 for cluster 0, 1 for cluster 1, and so on.}

\item{benjamini}{A number that represents the Benjamini threshold for
filtering processes. Only processes with a Benjamini value smaller than the
threshold will be kept. The default value is 0.05.}

\item{top_n}{The number of processes to return.}
}
\value{
A data frame with the top_n top processes.
}
\description{
This function find the top GO processes identified in a cluster, based on
the Benjamini threshold given.
}
\examples{
\dontrun{
library(SeuratToGO)
combined_list <- combine_david_files("./david")
get_top_processes(combined_list, cluster = 0, benjamini = 0.05, top_n = 5)
}

}
\references{
Sherman BT, Hao M, Qiu J, Jiao X, Baseler MW, Lane HC, Imamichi T, Chang W.
2022. DAVID: a web server for functional enrichment analysis and functional
annotation of gene lists (2021 update). \emph{Nucleic Acids Res}.
50(W1):W216–W221. \href{https://pubmed.ncbi.nlm.nih.gov/35325185/}{Link}

Benjamini Y, Hochberg Y. 1995. Controlling the False Discovery Rate: A
Practical and Powerful Approach to Multiple Testing. \emph{Journal of the
Royal Statistical Society: Series B (Methodological).} 57(1):289–300.
\href{https://rss.onlinelibrary.wiley.com/doi/10.1111/j.2517-6161.1995.tb02031.x}{Link}
}
