% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_processes_heatmap.R
\name{top_processes_heatmap}
\alias{top_processes_heatmap}
\title{Produce a heatmap of the top processes}
\usage{
top_processes_heatmap(
  top_processes_df,
  width = 12,
  height = 6,
  cellwidth = 30,
  save = FALSE
)
}
\arguments{
\item{top_processes_df}{A data frame containing the top processes of each
cluster. The rownames should be the processes and the column names should
be the cluster. It can be produced using the get_all_top_processes function
in the package.}

\item{width}{The width of the output file}

\item{height}{The height of the output file}

\item{cellwidth}{The cell width for each cell in the heatmap.}

\item{save}{TRUE if the user wants to save the plot as a png image file. The
default value is FALSE.}
}
\value{
A heatmap that represents the top processes in each cluster.
}
\description{
This function creates and returns a heatmap showing the top processes in each
cluster. It automatically saves the plot as a .png file in the current
working directory.
}
\examples{
\dontrun{
library(SeuratToGO)
david_file_path = system.file("extdata", "david", package = "SeuratToGO")
# this is where DAVID output files are stored in this package
# when using your own files, you must provide the file path to the folder
# where you saved them

combined_list <- combine_david_files(david_file_path)
top_processes <- get_all_top_processes(combined_list)
top_processes_heatmap(top_df)
# you should see a heat map in the plot window

top_processes_heatmap(top_df, save = TRUE)
# if you want to save the plot in your current directory, set save to TRUE
}
}
